+++ docs/NEWS.rst
@@ -8,7 +8,7 @@ of XIST. For a description of how to update your code to each versions of XIST
 see :ref:`MIGRATION`.
 
 
-Changes in 5.78 (released 2025-??-??)
+Changes in 5.78 (released 2025-03-17)
 -------------------------------------
 
 *	CSS output of colors with an alpha channel now no longer uses the